# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Participedia
  description: This API is the proxy used by Participedia.net to talk to the database of records in the database, as well as the user database.  It should be considered the stable interface to the data.  Some data sources may have their own API, but those are subject to change without notice.
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  token:
    type: apiKey
    description: For all of the API calls below, any call that causes a change in state on the server must be signed with a JWT token provided by Auth0.  The credentials of the user will be used to determine whether the user has the permissions required to perform the requested operation.  See Auth0 docs for details.
    name:
      Authorization
    in:
      header
paths:
...
  /case/new:
    x-swagger-router-controller: case
    put:
      description: Submit a new case
      operationId: newCase
      tags:
        - case
      produces:
        - application/json
      security:
        - token: []
      parameters:
        - name: case
          in: body
          description: the new case data
          schema:
            $ref: '#/definitions/case'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/case'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /case/{caseId}:
    x-swagger-router-controller: case
    put:
      description: Submit a new revision of a case
      operationId: editCaseById
      tags:
        - case
      produces:
        - application/json
      security:
        - token: []
      parameters:
        - name: caseId
          in: path
          description: the caseId
          type: integer
          required: true
        - name: case
          in: body
          description: the new version of the case
          required: true
          schema:
            $ref: '#/definitions/case'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/case'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    get:
      description: Get a case
      operationId: getCaseById
      tags:
        - case
      produces:
        - application/json
      security:
        - token: []
      parameters:
        - name: caseId
          in: path
          description: the caseId
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/case'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    delete:
      description: Delete a case
      operationId: deleteCaseById
      tags:
        - case
      security:
        - token: []
      parameters:
        - name: caseId
          in: path
          description: the caseId
          type: integer
          required: true
      responses:
        "200":
          description: Success
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
definitions:
  case:
    type: object
    required:
      - id
      - title
    properties:
      id:
        type: integer
      title:
        type: string
...
  errorModel:
      type: string
      required:
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string